% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimble_alpha.R
\docType{data}
\name{alphaGroupPostTermsListDef}
\alias{alphaGroupPostTermsListDef}
\title{Definition of nimbleList for grouped posterior terms}
\format{
An object of class \code{list} of length 1.
}
\usage{
alphaGroupPostTermsListDef
}
\description{
Internal container storing posterior precision and linear terms.
}
\keyword{internal}
