% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_loss}
\alias{compute_loss}
\title{Loss Function for Coupled Unmixing}
\usage{
compute_loss(A, B_list, X_list)
}
\arguments{
\item{A}{Matrix (N x K) - mixing proportions}

\item{B_list}{List of source signature matrices}

\item{X_list}{List of data matrices}
}
\value{
Scalar loss value (sum of squared errors)
}
\description{
Computes total reconstruction error across all data blocks.
}
\keyword{internal}
