% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_weighted_loss}
\alias{compute_weighted_loss}
\title{Weighted Loss Function}
\usage{
compute_weighted_loss(A, B_list, X_list, W, global_weights)
}
\arguments{
\item{A}{Matrix (N x K) - mixing proportions}

\item{B_list}{List of source signature matrices}

\item{X_list}{List of data matrices}

\item{W}{Matrix of sample weights (N x n_blocks)}

\item{global_weights}{Named vector of global block weights}
}
\value{
Scalar loss value
}
\description{
Computes weighted reconstruction error across all data blocks.
}
\keyword{internal}
