% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaprelims.R
\docType{package}
\name{geomix-package}
\alias{geomix}
\alias{geomix-package}
\title{geomix: GeoMix: Bayesian Spatial Mixture Modelling for Offshore Site Characterisation}
\description{
Tools for fitting GeoMix models for spatial classification and prediction using Bayesian hierarchical mixture models. The package combines Potts latent class models, grouped Vecchia Gaussian process approximations, custom NIMBLE samplers, posterior prediction, and MCMC diagnostic utilities. Developed for offshore ground characterisation using geophysical and geotechnical data.
}
\author{
\strong{Maintainer}: Bradley Wakefield \email{bradleyw@uow.edu.au}

}
\keyword{internal}
