% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hellinger_distance}
\alias{hellinger_distance}
\title{Hellinger Distance}
\usage{
hellinger_distance(p, q)
}
\arguments{
\item{p}{First distribution (will be normalized)}

\item{q}{Second distribution (will be normalized)}
}
\value{
Hellinger distance (0 to 1)
}
\description{
Computes Hellinger distance between two distributions.
}
