% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_optim.R
\name{k_find_elbow}
\alias{k_find_elbow}
\title{Detect an "elbow" in the K–loss curve}
\usage{
k_find_elbow(x, y)
}
\arguments{
\item{x}{Numeric vector of K values (increasing order).}

\item{y}{Numeric vector of loss values.}
}
\value{
Integer K corresponding to the elbow (or `NA`).
}
\description{
Simple geometric heuristic (distance to diagonal method).
}
