% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nnls_A_rows}
\alias{nnls_A_rows}
\title{NNLS for A Given B (Row-wise)}
\usage{
nnls_A_rows(B, X)
}
\arguments{
\item{B}{Matrix (K x F) - source signatures}

\item{X}{Matrix (N x F) - data}
}
\value{
Matrix A (N x K) - mixing proportions
}
\description{
Solves for A in X ~ A %*% B using non-negative least squares,
row by row.
}
\keyword{internal}
