% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nnls_B_given_A}
\alias{nnls_B_given_A}
\title{NNLS for B Given A}
\usage{
nnls_B_given_A(A, X)
}
\arguments{
\item{A}{Matrix (N x K) - mixing proportions}

\item{X}{Matrix (N x F) - data}
}
\value{
Matrix B (K x F) - source signatures
}
\description{
Solves for B in X ~ A %*% B using non-negative least squares,
column by column.
}
\keyword{internal}
