% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_block_by_col}
\alias{normalize_block_by_col}
\title{Normalize Data Block by Column}
\usage{
normalize_block_by_col(B, X)
}
\arguments{
\item{B}{Matrix (K x F) - source signatures}

\item{X}{Matrix (N x F) - data}
}
\value{
List with normalized Bn and Xn
}
\description{
Normalizes both B (source signatures) and X (data) by column-wise scaling.
This ensures fair weighting between different data blocks.
}
\keyword{internal}
