% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ols_B_given_A}
\alias{ols_B_given_A}
\title{OLS for B Given A}
\usage{
ols_B_given_A(A, X)
}
\arguments{
\item{A}{Matrix (N x K) - mixing proportions}

\item{X}{Matrix (N x F) - data}
}
\value{
Matrix B (K x F) - source signatures
}
\description{
Solves for B in X ~ A %*% B using ordinary least squares.
}
\keyword{internal}
