% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.unmix_result}
\alias{plot.unmix_result}
\title{Plot Methods for Unmixing Results}
\usage{
\method{plot}{unmix_result}(x, type = c("mixing", "sources", "loss"), ...)
}
\arguments{
\item{x}{An unmix_result or bootstrap_result object}

\item{type}{Type of plot: "mixing" (default), "sources", "loss", "uncertainty"}

\item{...}{Additional plotting parameters}
}
\description{
Visualize unmixing results including mixing proportions,
source signatures, and uncertainty estimates.
}
