% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_loss_history}
\alias{plot_loss_history}
\title{Plot Loss History}
\usage{
plot_loss_history(x, log = TRUE, ...)
}
\arguments{
\item{x}{An unmix_result object}

\item{log}{Logical, use log scale for y-axis? (default: TRUE)}

\item{...}{Additional parameters}
}
\description{
Plot Loss History
}
\keyword{internal}
