% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_mixing_matrix}
\alias{plot_mixing_matrix}
\title{Plot Mixing Matrix}
\usage{
plot_mixing_matrix(x, samples = NULL, ...)
}
\arguments{
\item{x}{A provunmix object}

\item{samples}{Optional vector of sample indices to plot (default: all)}

\item{...}{Additional parameters}
}
\description{
Plot Mixing Matrix
}
\keyword{internal}
