% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sources}
\alias{plot_sources}
\title{Plot Source Signatures}
\usage{
plot_sources(x, block = 1, ...)
}
\arguments{
\item{x}{An unmix_result object}

\item{block}{Which data block to plot (default: 1)}

\item{...}{Additional parameters}
}
\description{
Plot Source Signatures
}
\keyword{internal}
