% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_uncertainty}
\alias{plot_uncertainty}
\title{Plot Uncertainty Estimates}
\usage{
plot_uncertainty(x, samples = NULL, ...)
}
\arguments{
\item{x}{A provunmix_boot object}

\item{samples}{Optional vector of sample indices to plot (default: first 10)}

\item{...}{Additional parameters}
}
\description{
Plot Uncertainty Estimates
}
\keyword{internal}
