% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimble_groupVecchia.R
\docType{data}
\name{predictOutputListDef}
\alias{predictOutputListDef}
\title{Definition of nimbleList for Gaussian process prediction outputs}
\format{
An object of class \code{list} of length 1.
}
\usage{
predictOutputListDef
}
\description{
Internal container storing predictive means, standard deviations, and
covariance matrices for grouped Gaussian process predictions.
}
\keyword{internal}
