% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{print.bootstrap_result}
\alias{print.bootstrap_result}
\title{Print Method for bootstrap_result Objects}
\usage{
\method{print}{bootstrap_result}(x, ...)
}
\arguments{
\item{x}{A bootstrap_result object}

\item{...}{Additional arguments (unused)}
}
\description{
Print Method for bootstrap_result Objects
}
