% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance_unmix.R
\name{print.provunmix}
\alias{print.provunmix}
\title{Print Method for provunmix Objects}
\usage{
\method{print}{provunmix}(x, ...)
}
\arguments{
\item{x}{A provunmix object}

\item{...}{Additional arguments (unused)}
}
\description{
Print Method for provunmix Objects
}
