% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{print.provunmix_boot}
\alias{print.provunmix_boot}
\title{Print Method for provunmix_boot Objects}
\usage{
\method{print}{provunmix_boot}(x, ...)
}
\arguments{
\item{x}{A provunmix_boot object}

\item{...}{Additional arguments (unused)}
}
\description{
Print Method for provunmix_boot Objects
}
