% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance_unmix.R
\name{print.unmix_result}
\alias{print.unmix_result}
\title{Print Method for unmix_result Objects}
\usage{
\method{print}{unmix_result}(x, ...)
}
\arguments{
\item{x}{An unmix_result object}

\item{...}{Additional arguments (unused)}
}
\description{
Print Method for unmix_result Objects
}
