% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{summary.bootstrap_result}
\alias{summary.bootstrap_result}
\title{Summary Method for bootstrap_result Objects}
\usage{
\method{summary}{bootstrap_result}(object, ...)
}
\arguments{
\item{object}{A bootstrap_result object}

\item{...}{Additional arguments (unused)}
}
\description{
Summary Method for bootstrap_result Objects
}
