% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance_unmix.R
\name{summary.provunmix}
\alias{summary.provunmix}
\title{Summary Method for provunmix Objects}
\usage{
\method{summary}{provunmix}(object, ...)
}
\arguments{
\item{object}{A provunmix object}

\item{...}{Additional arguments (unused)}
}
\description{
Summary Method for provunmix Objects
}
