% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{summary.provunmix_boot}
\alias{summary.provunmix_boot}
\title{Summary Method for provunmix_boot Objects}
\usage{
\method{summary}{provunmix_boot}(object, ...)
}
\arguments{
\item{object}{A provunmix_boot object}

\item{...}{Additional arguments (unused)}
}
\description{
Summary Method for provunmix_boot Objects
}
