% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R, R/utils.R
\name{summary.unmix_cv_result}
\alias{summary.unmix_cv_result}
\title{Summary Method for unmix_cv_result}
\usage{
\method{summary}{unmix_cv_result}(object, ...)

\method{summary}{unmix_cv_result}(object, ...)
}
\description{
Summary Method for unmix_cv_result

Summary Method for unmix_cv_result
}
