% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance_unmix.R
\name{summary.unmix_result}
\alias{summary.unmix_result}
\title{Summary Method for unmix_result Objects}
\usage{
\method{summary}{unmix_result}(object, ...)
}
\arguments{
\item{object}{An unmix_result object}

\item{...}{Additional arguments (unused)}
}
\description{
Summary Method for unmix_result Objects
}
