% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{summary.unmix_cv_result}
\alias{summary.unmix_cv_result}
\title{Summary Method for unmix_cv_result}
\usage{
\method{summary}{unmix_cv_result}(object, ...)
}
\arguments{
\item{object}{An unmix_cv_result object.}

\item{...}{Additional arguments (unused).}
}
\description{
Summary Method for unmix_cv_result
}
